Add option to build wheels with locked dependencies
By default, pyproject dependencies have priority over those
in the lock file. For example, if 'pyproject.toml' requires
'pandas = ">=1.3' and 'poetry.lock' requires 'pandas 1.4.2',
the wheel file will contain the next line for this dependency:

  Requires-Dist: pandas (>=0.1.3,<0.2.0)

With the new option 'only-lock', the wheel file will only
include those fixed dependencies listed on the lock file
and not on the 'pyproject.toml', so the result for the
previous example will be:

  Requires-Dist: pandas (==1.4.2)

This will make easier to have deterministic environments,
where every package installed is matches with the fixed
versions of the lock file.

However, poetry requires that the packages listed on the
'pyproject.toml' are also on the 'poetry.lock' file,
so no direct dependencies will be missed when the wheel
is built. Only package versions may vary.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>